Changed incrementing `pos` in `((left operand[pos]) | (right operand[++pos]))` to just incrementing in a newline for `base64_decode`
Although it's inefficient, Some compilers seem to (`MSVC` on windows) seems to compile from right to left (Not done much research on this) so it increments the value on right operand `((left operand[pos]) | (right operand[++pos]))` first unlike `gcc` (Tested with `gcc` on Linux) which seems to go from left to right executing the left operand first.

This causes `base64_decode` to basically not work, Some random bytes are just returned as the algorithm logic breaks because it expects the left operand to be executed first.